Expired users

+++ src/controllers/users.controller.js
@@ -6,7 +6,7 @@ import { userServices } from "../services/service.js";
 import jwt from "jsonwebtoken";
 import { v2 } from "../config/config.js";
 import { dataUri } from "../utils.js";
-import { sendEmailToDeletedUser } from "../utils/sendEmailToDeletedUser.js";
+import { sendEmailToDeletedUser } from "../utils/sendEmail.js";
 
 //Register
 export const jwtRegisterController = (req, res, next) => {
@@ -235,24 +235,31 @@ export const getAllUsersController = async (req, res) => {
 export const deleteExpiredCountsController = async (req, res) => {
   try {
     const users = await userServices.getUsers();
+    const expiredUser = [];
     for (const user of users) {
-      const expirationTime = user.last_connection + /* 48 * 60 */ 1 * 60 * 1000;
+      const expirationTime =
+        user.last_connection +  48 * 60 * 60 * 1000;
       const timeNumber = Date.now();
       if (
-        expirationTime < timeNumber ||
-        user.last_connection == null ||
-        user.last_connection == undefined
+        (expirationTime < timeNumber ||
+          user.last_connection == null ||
+          user.last_connection == undefined) &&
+        user.role !== "Admin"
       ) {
-        console.log("Borrar");
         await userServices.deleteUser(user._id);
         await sendEmailToDeletedUser(user);
+        expiredUser.push(user);
       }
     }
-    res
-      .status(200)
-      .send({
-        message: "Usuario inactivos durante 48hs. eliminados exitosamente.",
-      });
+    if (expiredUser.length === 0) {
+      return res
+        .status(204)
+        .send({ message: "Todos los usuarios se encuentran activos" });
+    }
+    res.status(200).send({
+      message: "Usuario inactivos durante 48hs. eliminados exitosamente.",
+      expiredUser,
+    });
   } catch (error) {
     res.status(404).json({
       message: error.message,
